Bug fix
The previous format of the sale id, did not allow to retrieve the mentioned id in exadecimal format via get request. By setting it to String the get request now returns the correct value instead of an object in {date:, timestamp:} format thus allowing to delete end edit the record on the front-end side.